fix: split knowledge corpus security boundaries
