+++ todo/tasks/28591-brief.md
@@ -91,7 +91,9 @@ the parent; successful merge should trigger Phase 2 auto-filing.
 ### Files to Modify
 
 - `NEW: .agents/scripts/knowledge-corpus-helper.sh` — narrow shell CLI and environment contract.
-- `NEW: .agents/scripts/knowledge_corpus_helpers.py` — SQLite catalog, local principal context, schema bootstrap, and authorized resolution.
+- `NEW: .agents/scripts/knowledge_corpus_helpers.py` — narrow Python argument parsing and command dispatch.
+- `NEW: .agents/scripts/knowledge_corpus_catalog.py` — SQLite schema, bootstrap graph, and authorized catalog queries.
+- `NEW: .agents/scripts/knowledge_corpus_context.py` — owner-only context and filesystem/path validation primitives.
 - `EDIT: .agents/scripts/knowledge-helper.sh` — provision the catalog in personal mode and route personal add/list/search through capability-aware alias resolution.
 - `NEW: .agents/tests/test-knowledge-corpus.sh` — schema, idempotency, permissions, authorization, isolation, and unsafe-path tests.
 - `EDIT: .agents/tests/test-knowledge-cli.sh` — isolate the Vault fixture so the compatibility gate cannot inherit a locked user Vault.
@@ -101,8 +103,8 @@ the parent; successful merge should trigger Phase 2 auto-filing.
 
 ### Complete Write Surface
 
-- **Callers/readers:** `.agents/scripts/knowledge-helper.sh` calls `.agents/scripts/knowledge-corpus-helper.sh`; the shell wrapper invokes `.agents/scripts/knowledge_corpus_helpers.py`; users call the existing `aidevops knowledge` surface.
-- **Writers/mutation paths:** `.agents/scripts/knowledge_corpus_helpers.py` atomically creates the local authentication context and writes `catalog.db`; `.agents/scripts/knowledge-helper.sh` continues writing the unchanged legacy `_knowledge` tree.
+- **Callers/readers:** `.agents/scripts/knowledge-helper.sh` calls `.agents/scripts/knowledge-corpus-helper.sh`; the shell wrapper invokes `.agents/scripts/knowledge_corpus_helpers.py`, which delegates catalog work to `.agents/scripts/knowledge_corpus_catalog.py` and private-path/context checks to `.agents/scripts/knowledge_corpus_context.py`; users call the existing `aidevops knowledge` surface.
+- **Writers/mutation paths:** `.agents/scripts/knowledge_corpus_catalog.py` writes `catalog.db`; `.agents/scripts/knowledge_corpus_context.py` atomically creates the local authentication context; `.agents/scripts/knowledge-helper.sh` continues writing the unchanged legacy `_knowledge` tree.
 - **Tests/fixtures:** `NEW: .agents/tests/test-knowledge-corpus.sh`; `EDIT: tests/test-knowledge-provisioning.sh`; `EDIT: .agents/tests/test-knowledge-cli.sh` to isolate its Vault fixture after a locked user Vault proved the raw gate non-hermetic.
 - **Schemas/config:** `NEW: catalog.db` schema v1 with the parent catalog tables plus private `corpus_aliases`, and owner-only `_config/principal.json`; `repos.json` values and `.agents/templates/knowledge-config.json` remain unchanged.
 - **Generated/deployed mirrors:** No committed mirror is edited because `setup.sh` deploys tracked `.agents/scripts/` files; verify mirror/deployment coverage through existing changed-file gates.
@@ -111,7 +113,7 @@ the parent; successful merge should trigger Phase 2 auto-filing.
 
 ### Implementation Steps
 
-1. Add schema-v1 initialization with the parent catalog tables, private alias indirection, WAL, foreign keys, busy timeout, atomic transactions, and `0600` database/context files beneath `0700` private directories.
+1. Add bounded CLI, catalog, and context modules implementing schema-v1 initialization with the parent catalog tables, private alias indirection, WAL, foreign keys, busy timeout, atomic transactions, and `0600` database/context files beneath `0700` private directories.
 2. Bootstrap one opaque local principal, personal workspace, corpus, membership, and `knowledge.read|write|manage` grants; persist only the opaque principal ID in the authentication context.
 3. Register `personal:default` with `location_ref=<personal-base>/_knowledge`; reject conflicting aliases or out-of-base locations.
 4. Add list/resolve commands that derive the principal from the validated context file and require every active authorization edge before returning a path.
@@ -134,7 +136,7 @@ bash .agents/tests/test-knowledge-corpus.sh
 bash tests/test-knowledge-provisioning.sh
 bash .agents/tests/test-knowledge-cli.sh
 shellcheck .agents/scripts/knowledge-corpus-helper.sh .agents/scripts/knowledge-helper.sh .agents/tests/test-knowledge-corpus.sh tests/test-knowledge-provisioning.sh
-python3 -m py_compile .agents/scripts/knowledge_corpus_helpers.py
+python3 -m py_compile .agents/scripts/knowledge_corpus_helpers.py .agents/scripts/knowledge_corpus_catalog.py .agents/scripts/knowledge_corpus_context.py
 .agents/scripts/linters-local.sh --changed
 ```
 
@@ -163,6 +165,8 @@ python3 -m py_compile .agents/scripts/knowledge_corpus_helpers.py
 
 - `.agents/scripts/knowledge-corpus-helper.sh`
 - `.agents/scripts/knowledge_corpus_helpers.py`
+- `.agents/scripts/knowledge_corpus_catalog.py`
+- `.agents/scripts/knowledge_corpus_context.py`
 - `.agents/scripts/knowledge-helper.sh`
 - `.agents/tests/test-knowledge-corpus.sh`
 - `.agents/tests/test-knowledge-cli.sh` (edited after the compatibility gate inherited a locked user Vault)
@@ -202,7 +206,7 @@ python3 -m py_compile .agents/scripts/knowledge_corpus_helpers.py
   ```yaml
   verify:
     method: bash
-    run: "shellcheck .agents/scripts/knowledge-corpus-helper.sh .agents/scripts/knowledge-helper.sh .agents/tests/test-knowledge-corpus.sh tests/test-knowledge-provisioning.sh && python3 -m py_compile .agents/scripts/knowledge_corpus_helpers.py && .agents/scripts/linters-local.sh --changed"
+    run: "shellcheck .agents/scripts/knowledge-corpus-helper.sh .agents/scripts/knowledge-helper.sh .agents/tests/test-knowledge-corpus.sh tests/test-knowledge-provisioning.sh && python3 -m py_compile .agents/scripts/knowledge_corpus_helpers.py .agents/scripts/knowledge_corpus_catalog.py .agents/scripts/knowledge_corpus_context.py && .agents/scripts/linters-local.sh --changed"
   ```
 
 ## Context & Decisions
@@ -211,6 +215,7 @@ python3 -m py_compile .agents/scripts/knowledge_corpus_helpers.py
 - Catalog IDs and corpus-directory IDs are random opaque values. `personal:default` is a private logical alias, not a physical path component.
 - Explicit capability grants are retained even for personal owner access so later workspace support does not require a permissive compatibility exception.
 - Repo-mode knowledge remains outside the personal/workspace catalog in this phase.
+- CLI, catalog, and context responsibilities are separate modules so each security boundary stays reviewable and the Qlty new-file complexity gate remains clean.
 
 ## Relevant Files
 
@@ -220,6 +225,8 @@ python3 -m py_compile .agents/scripts/knowledge_corpus_helpers.py
 - `.agents/scripts/knowledge-helper.sh:223-277` — current provisioning path.
 - `.agents/scripts/knowledge-helper.sh:430-452` — current mode-to-root resolver.
 - `.agents/scripts/knowledge-helper.sh:1006-1097` — current search path.
+- `.agents/scripts/knowledge_corpus_catalog.py:1-343` — catalog schema, transactional bootstrap, and authorization query.
+- `.agents/scripts/knowledge_corpus_context.py:1-188` — private path, file, and authenticated-context validation.
 - `.agents/aidevops/knowledge-plane/01-core-contract.md:14-40` — current modes/layout.
 - `.agents/tools/app-stack/workspace-model.md:45-67` — workspace boundary rules.
 - `.agents/tools/app-stack/rbac-permissions.md:38-67` — default-deny membership/capability evaluation.
